requirements.txt
streamlit
pandas
xlsxwriter
google-generativeai